update-tugas

+++ tugas_thread_server.py
@@ -18,28 +18,32 @@ def handle_thread(conn):
             #terima string dari client
             data = conn.recv(100)
             data = data.decode('ascii')
-
             datasp = data.split()
-            if datasp[0] == "new":
-                files = open(datasp[1],"w")
-                res = data.split(datasp[1]+" ")
-                files.write(res[1])
-                respon = "OK"
-                files.close()
-            elif datasp[0] == "read":
-                files = open(datasp[1],"r")
-                respon = files.read()
-                files.close()
-            elif datasp[0] == "del":
-                os.remove(datasp[1])
-                respon = "OK"
-            else:
-                respon = "perintah tidak ditemukan"
+            if data == "exit":
+                #tutup koneksi
+                print("koneksi dituup")
+                break
+            else:    
+                if datasp[0] == "new":
+                    files = open(datasp[1],"w")
+                    res = data.split(datasp[1]+" ")
+                    files.write(res[1])
+                    respon = "OK"
+                    files.close()
+                elif datasp[0] == "read":
+                    files = open(datasp[1],"r")
+                    respon = files.read()
+                    files.close()
+                elif datasp[0] == "del":
+                    os.remove(datasp[1])
+                    respon = "OK"
+                else:
+                    respon = "perintah tidak ditemukan"
 
-            #kirim balik respon
-            conn.send(respon.encode('ascii'))
+                #kirim balik respon
+                conn.send(respon.encode('ascii'))
         except(socket.error):
-        	#tutup koneksi
+            #tutup koneksi
             print("koneksi dituup")
             break
             #conn.close()
